--- bucket-api/src/main/java/org/ecommerce/bucketapi/dto/BucketMapper.java
@@ -1,14 +1,0 @@
-package org.ecommerce.bucketapi.dto;
-
-import org.ecommerce.bucketapi.entity.Bucket;
-import org.mapstruct.Mapper;
-import org.mapstruct.ReportingPolicy;
-import org.mapstruct.factory.Mappers;
-
-@Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
-public interface BucketMapper {
-
-	BucketMapper INSTANCE = Mappers.getMapper(BucketMapper.class);
-	BucketDto toDto(Bucket bucket);
-	BucketDto.Response toResponse(BucketDto bucketDto);
-}
