--- bucket-api/src/main/resources/application-local.yml
@@ -1,28 +1,0 @@
-server:
-  port: 8080
-spring:
-  config:
-    activate:
-      on-profile: local
-  application:
-    name: bucket-api
-  h2:
-    console:
-      enabled: true
-      settings:
-        web-allow-others: true
-      path: /h2-console
-  jpa:
-    hibernate:
-      ddl-auto: create-drop
-    show-sql: true
-    generate-ddl: true
-    defer-datasource-initialization: true
-  sql:
-    init:
-      mode: always
-  datasource:
-    driver-class-name: org.h2.Driver
-    url: jdbc:h2:mem:test
-    username: sa
-    password: 1234
