--- bucket-api/src/main/resources/application.yml
@@ -1,3 +1,0 @@
-spring:
-  profiles:
-    active: local
