--- bucket-api/src/main/resources/data.sql
@@ -1,2 +1,0 @@
-insert into bucket(id, user_id, seller, product_id, quantity)
-    values (1, 1, 'seller1',1, 7);
