--- bucket-api/src/test/java/org/ecommerce/bucketapi/BucketApiApplicationTests.java
@@ -1,13 +1,0 @@
-package org.ecommerce.bucketapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest(classes = BucketApiApplication.class)
-class BucketApiApplicationTests {
-
-  @Test
-  void contextLoads() {
-  }
-
-}
