+++ order-api/src/main/java/org/ecommerce/orderapi/bucket/dto/BucketDto.java
@@ -1,0 +1,13 @@
+package org.ecommerce.orderapi.bucket.dto;
+
+import java.time.LocalDate;
+
+public record BucketDto(
+		Long id,
+		Integer userId,
+		String seller,
+		Integer productId,
+		Integer quantity,
+		LocalDate createDate
+) {
+}
