+++ order-api/src/main/java/org/ecommerce/orderapi/bucket/dto/response/BucketResponse.java
@@ -1,0 +1,13 @@
+package org.ecommerce.orderapi.bucket.dto.response;
+
+import java.time.LocalDate;
+
+public record BucketResponse(
+		Long id,
+		Integer userId,
+		String seller,
+		Integer productId,
+		Integer quantity,
+		LocalDate createDate
+) {
+}
