+++ order-api/src/main/java/org/ecommerce/orderapi/bucket/exception/ErrorMessage.java
@@ -1,0 +1,8 @@
+package org.ecommerce.orderapi.bucket.exception;
+
+public class ErrorMessage {
+	public static final String ERROR_SELLER_REQUIRED = "판매자를 입력해 주세요.";
+	public static final String ERROR_PRODUCT_ID_REQUIRED = "상품 번호를 입력해 주세요.";
+	public static final String ERROR_QUANTITY_REQUIRED = "상품 수량을 입력해 주세요.";
+	public static final String ERROR_QUANTITY_MIN = "상품 수량을 1개 이상으로 입력해 주세요.";
+}
