--- order-api/src/main/java/org/ecommerce/orderapi/order/dto/response/BucketResponse.java
@@ -1,13 +1,0 @@
-package org.ecommerce.orderapi.order.dto.response;
-
-import java.time.LocalDate;
-
-public record BucketResponse(
-		Long id,
-		Integer userId,
-		String seller,
-		Integer productId,
-		Integer quantity,
-		LocalDate createDate
-) {
-}
