--- order-api/src/main/java/org/ecommerce/orderapi/order/entity/Bucket.java
@@ -1,27 +1,0 @@
-package org.ecommerce.orderapi.order.entity;
-
-import java.time.LocalDate;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-public class Bucket {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-	private Integer userId;
-	private String seller;
-	private Integer productId;
-	private Integer quantity;
-	private LocalDate createDate;
-}
