+++ order-api/src/main/java/org/ecommerce/orderapi/order/service/OrderDomainService.java
@@ -10,8 +10,8 @@ import org.ecommerce.common.error.CustomException;
 import org.ecommerce.orderapi.order.client.BucketServiceClient;
 import org.ecommerce.orderapi.order.client.PaymentServiceClient;
 import org.ecommerce.orderapi.order.client.ProductServiceClient;
-import org.ecommerce.orderapi.order.dto.BucketMapper;
-import org.ecommerce.orderapi.order.dto.BucketSummary;
+import org.ecommerce.orderapi.bucket.dto.BucketMapper;
+import org.ecommerce.orderapi.bucket.dto.BucketSummary;
 import org.ecommerce.orderapi.order.dto.OrderDtoWithOrderItemDtoList;
 import org.ecommerce.orderapi.order.dto.OrderMapper;
 import org.ecommerce.orderapi.order.dto.PaymentMapper;
