+++ settings.gradle
@@ -5,6 +5,5 @@ include 'payment-api'
 include 'product-management-api'
 include 'product-search-api'
 include 'statistic-api'
-include 'bucket-api'
 include 'common'
 include 'product'
